Remade Navigation Panel so it utilize subpage navigation system.
