--- StudyConfigurationUI/StudyConfigurationUI/View/Pages/PhaseCreationPages/PhaseTeamRolesPage.xaml
@@ -1,99 +1,0 @@
-﻿<Page
-    x:Class="StudyConfigurationUI.View.Pages.PhaseCreationPages.PhaseTeamRolesPage"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:StudyConfigurationUI.View.Pages.PhaseCreationPages"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    mc:Ignorable="d">
-
-
-    <Grid Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
-        <Grid.RowDefinitions>
-            <RowDefinition Height="Auto" />
-            <RowDefinition Height="*" />
-        </Grid.RowDefinitions>
-        <!--Adaptive Triggers-->
-        <VisualStateManager.VisualStateGroups>
-            <VisualStateGroup>
-                <VisualState x:Name="NarrowState">
-                    <VisualState.StateTriggers>
-                        <AdaptiveTrigger MinWindowWidth="0" />
-                    </VisualState.StateTriggers>
-                    <VisualState.Setters>
-                        <Setter Target="OverviewLabel.FontSize" Value="12" />
-                        <Setter Target="TeamsLabel.FontSize" Value="12" />
-                        <Setter Target="TasksLabel.FontSize" Value="12" />
-                        <Setter Target="CriteriasLabel.FontSize" Value="12" />
-                    </VisualState.Setters>
-                </VisualState>
-
-                <VisualState x:Name="CompactState">
-                    <VisualState.StateTriggers>
-                        <AdaptiveTrigger MinWindowWidth="720" />
-                    </VisualState.StateTriggers>
-                    <VisualState.Setters>
-                        <Setter Target="OverviewLabel.FontSize" Value="13" />
-                        <Setter Target="TeamsLabel.FontSize" Value="13" />
-                        <Setter Target="TasksLabel.FontSize" Value="13" />
-                        <Setter Target="CriteriasLabel.FontSize" Value="13" />
-                    </VisualState.Setters>
-                </VisualState>
-
-                <VisualState x:Name="WideState">
-                    <VisualState.StateTriggers>
-                        <AdaptiveTrigger MinWindowWidth="1080" />
-                    </VisualState.StateTriggers>
-                    <VisualState.Setters>
-                        <Setter Target="OverviewLabel.FontSize" Value="14" />
-                        <Setter Target="TeamsLabel.FontSize" Value="14" />
-                        <Setter Target="TasksLabel.FontSize" Value="14" />
-                        <Setter Target="CriteriasLabel.FontSize" Value="14" />
-                    </VisualState.Setters>
-                </VisualState>
-            </VisualStateGroup>
-
-        </VisualStateManager.VisualStateGroups>
-
-        <!-- Top Navigation Bar DO NOT DELETE-->
-        <ListView Grid.Row="0"
-                  Height="40"
-                  VerticalAlignment="Center"
-                  ScrollViewer.HorizontalScrollBarVisibility="Hidden"
-                  ScrollViewer.VerticalScrollBarVisibility="Disabled"
-                  ScrollViewer.HorizontalScrollMode="Enabled"
-                  ScrollViewer.VerticalScrollMode="Disabled"
-                  ScrollViewer.ZoomMode="Disabled"
-                  SelectionMode="Single"
-                  Background="LightGray"
-                  SelectionChanged="MenuListBox_OnSelectionChanged">
-            <!-- Setting list view for horizontal display -->
-            <ListView.ItemsPanel>
-                <ItemsPanelTemplate>
-                    <StackPanel Orientation="Horizontal" />
-                </ItemsPanelTemplate>
-
-            </ListView.ItemsPanel>
-            <ListViewItem x:Name="OverviewBut">
-                <TextBlock x:Name="OverviewLabel" Text="Overview" FontSize="12" />
-            </ListViewItem>
-
-            <ListViewItem x:Name="TeamsBut" IsSelected="True">
-                <TextBlock x:Name="TeamsLabel" Text="Teams and Roles" FontSize="12" />
-            </ListViewItem>
-
-            <ListViewItem x:Name="TasksBut">
-                <TextBlock x:Name="TasksLabel" Text="Tasks" FontSize="12" />
-            </ListViewItem>
-
-            <ListViewItem x:Name="CriteriasBut">
-                <TextBlock x:Name="CriteriasLabel" Text="Criterias" FontSize="12" />
-            </ListViewItem>
-
-        </ListView>
-        <!--=================================-->
-        
-        
-        
-    </Grid>
-</Page>
