+++ StudyConfigurationUI/StudyConfigurationUI/View/Pages/PhaseCreationPages/SubPages/PhaseOverviewPage.xaml
@@ -1,0 +1,13 @@
+﻿<Page
+    x:Class="StudyConfigurationUI.View.Pages.PhaseCreationPages.SubPages.PhaseOverviewPage"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:local="using:StudyConfigurationUI.View.Pages.PhaseCreationPages.SubPages"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    mc:Ignorable="d">
+
+    <Grid Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
+        <TextBlock Text="OverviewPage" HorizontalAlignment="Center" VerticalAlignment="Center"/>
+    </Grid>
+</Page>
